// src/utils/Types.ts
export interface Anime {
  id: number;
  name: string;
  russian: string;
  image: {
    original: string;
    preview: string;
    x96: string;
    x48: string;
  };
  url: string;
  kind: string;
  score: number;
  status: string;
  episodes: number;
  episodes_aired: number;
  aired_on: string;
  released_on: string;
  rating: string;
  english: string[];
  japanese: string[];
  synonyms: string[];
  license_name_ru: string | null;
  duration: number;
  description: string;
  description_html: string;
  description_source: string | null;
  franchise: string;
  favoured: boolean;
  anons: boolean;
  ongoing: boolean;
  thread_id: number;
  topic_id: number;
  myanimelist_id: number;
  rates_scores_stats: { name: number; value: number }[];
  rates_statuses_stats: { name: string; value: number }[];
  updated_at: string;
  next_episode_at: string | null;
  fansubbers: string[];
  fandubbers: string[];
  licensors: string[];
  genres: { id: number; name: string; russian: string; kind: string; entry_type: string }[];
  studios: { id: number; name: string; filtered_name: string; real: boolean; image: string }[];
  videos: { id: number; url: string; image_url: string; player_url: string; name: string; kind: string; hosting: string }[];
  screenshots: { original: string; preview: string }[];
  user_rate: null;
  [key: string]: any;
}